Set data types to a standard ulong for x64, added problems 11, 12 , 13 and improved 9

+++ Solutions/Problem1.cs
@@ -5,7 +5,7 @@ namespace ProjectEuler
     public partial class Solutions
     {
         [Arguments(3, 5, 1000), Benchmark]
-        public ulong Problem1(uint p1, uint p2, ulong p3)
+        public ulong Problem1(ulong p1, ulong p2, ulong p3)
         {
             ulong result, product = p1 * p2, limit = p3 - 1;
